requirements.txt
pytorch==1.7.0
pytorch-lightning==1.0.5
torchvision==0.8.1
numpy==1.19.4
matplotlib==3.3.1
seaborn==0.11.0
pandas==1.1.3
tqdm==4.50.2
yaml==0.2.5
scikit-learn==0.23.2
pillow==8.0.0
requests==2.24.0
